backend/requirements.txt
# ─── Web Framework ───────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ─── Database ────────────────────────────────────────────
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ─── Authentication ──────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ─── Configuration ───────────────────────────────────────
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ─── HTTP Client ─────────────────────────────────────────
httpx==0.26.0

# ─── WebSocket ───────────────────────────────────────────
websockets==12.0

# ─── Utilities ───────────────────────────────────────────
python-dateutil==2.8.2